requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-celery-beat==2.2.1
django-filter==2.4.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
future==0.18.2
gunicorn==19.9.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.1.0
mccabe==0.6.1
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
